refactor(ai-chat): extract shared ConversationHistoryTable
Pulls the duplicated column config, action menu, single-item dialogs,
and useConversations wiring out of ConversationHistoryPage and
DrawerConversationHistory into one shared component. Page keeps
useTable + multi-select + bulk-delete; drawer keeps its flex/scroll
wrapper. Net -300 lines.
