+++ plugins/ai-chat/src/components/ConversationHistory/ConversationHistoryPage.tsx
@@ -1,56 +1,15 @@
 import { Content } from '@backstage/core-components';
-import {
-  Button,
-  ButtonIcon,
-  Cell,
-  CellText,
-  ColumnConfig,
-  Menu,
-  MenuItem,
-  MenuSeparator,
-  MenuTrigger,
-  Table,
-  Text,
-  useTable,
-} from '@backstage/ui';
+import { Button, useTable } from '@backstage/ui';
 import { makeStyles } from '@material-ui/core';
-import {
-  RiDeleteBinLine,
-  RiEditLine,
-  RiMoreLine,
-  RiStarFill,
-  RiStarLine,
-} from '@remixicon/react';
+import { RiDeleteBinLine } from '@remixicon/react';
 import { useState } from 'react';
-import { useConversations } from '../../hooks/useConversations';
-import {
-  DeleteConversationDialog,
-  DeleteConversationsDialog,
-} from '../DeleteConversationDialog';
-import { RenameConversationDialog } from '../RenameConversationDialog';
-import type { ConversationApi, ConversationListItem } from '../../api';
-import { formatRelativeDate, getConversationTitle } from '../../utils';
 import type { Selection } from 'react-aria-components';
+import type { ConversationApi, ConversationListItem } from '../../api';
+import { useConversations } from '../../hooks/useConversations';
+import { ConversationHistoryTable } from '../ConversationHistoryTable';
+import { DeleteConversationsDialog } from '../DeleteConversationDialog';
 
 const useStyles = makeStyles(theme => ({
-  emptyState: {
-    textAlign: 'center' as const,
-    padding: theme.spacing(2),
-  },
-  starIcon: {
-    color: theme.palette.warning.main,
-  },
-  actionsContainer: {
-    display: 'flex',
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    height: '100%',
-  },
-  activeTitle: {
-    '& p[data-weight="regular"]': {
-      fontWeight: 'var(--bui-font-weight-bold)',
-    },
-  },
   selectionToolbar: {
     display: 'flex',
     alignItems: 'center',
@@ -72,21 +31,11 @@ export const ConversationHistoryPage = ({
 }: ConversationHistoryPageProps) => {
   const classes = useStyles();
   const [selectedKeys, setSelectedKeys] = useState<Selection>(new Set());
-  const [pendingDelete, setPendingDelete] =
-    useState<ConversationListItem | null>(null);
-  const [pendingRename, setPendingRename] =
-    useState<ConversationListItem | null>(null);
   const [pendingBulkDelete, setPendingBulkDelete] = useState<string[] | null>(
     null,
   );
-  const {
-    conversations,
-    loading,
-    deleteConversation,
-    deleteConversations,
-    toggleStar,
-    renameConversation,
-  } = useConversations(conversationApi);
+  const { conversations, deleteConversations } =
+    useConversations(conversationApi);
 
   const { tableProps } = useTable<ConversationListItem>({
     mode: 'complete',
@@ -97,81 +46,6 @@ export const ConversationHistoryPage = ({
     selectedKeys === 'all'
       ? conversations.map(c => c.id)
       : Array.from(selectedKeys, String);
-
-  const columnConfig: ColumnConfig<ConversationListItem>[] = [
-    {
-      id: 'title',
-      label: 'Title',
-      isRowHeader: true,
-      cell: item => (
-        <CellText
-          className={item.id === activeId ? classes.activeTitle : undefined}
-          title={getConversationTitle(item)}
-          description={formatRelativeDate(item.updatedAt)}
-        />
-      ),
-    },
-    {
-      id: 'starred',
-      label: '',
-      width: 32,
-      cell: item => (
-        <Cell>
-          {item.isStarred && (
-            <RiStarFill size={16} className={classes.starIcon} />
-          )}
-        </Cell>
-      ),
-    },
-    {
-      id: 'actions',
-      label: '',
-      width: 48,
-      cell: item => (
-        <Cell>
-          <div className={classes.actionsContainer}>
-            <MenuTrigger>
-              <ButtonIcon
-                icon={<RiMoreLine />}
-                aria-label="Actions"
-                variant="tertiary"
-                size="small"
-              />
-              <Menu>
-                <MenuItem
-                  onAction={() => toggleStar(item.id)}
-                  iconStart={
-                    item.isStarred ? (
-                      <RiStarLine size={16} />
-                    ) : (
-                      <RiStarFill size={16} />
-                    )
-                  }
-                  style={{ display: 'none' }}
-                >
-                  {item.isStarred ? 'Unstar' : 'Star'}
-                </MenuItem>
-                <MenuItem
-                  onAction={() => setPendingRename(item)}
-                  iconStart={<RiEditLine size={16} />}
-                >
-                  Rename
-                </MenuItem>
-                <MenuSeparator />
-                <MenuItem
-                  onAction={() => setPendingDelete(item)}
-                  iconStart={<RiDeleteBinLine size={16} />}
-                  color="danger"
-                >
-                  Delete
-                </MenuItem>
-              </Menu>
-            </MenuTrigger>
-          </div>
-        </Cell>
-      ),
-    },
-  ];
 
   return (
     <Content>
@@ -186,40 +60,17 @@ export const ConversationHistoryPage = ({
           {selectedIds.length > 0 ? 'Delete selected' : 'Delete'}
         </Button>
       </div>
-      <Table
-        {...tableProps}
-        columnConfig={columnConfig}
-        loading={loading || tableProps.loading}
+      <ConversationHistoryTable
+        conversationApi={conversationApi}
+        activeId={activeId}
+        onSelectConversation={onSelectConversation}
+        tableProps={tableProps}
         selection={{
           mode: 'multiple',
           behavior: 'toggle',
           selected: selectedKeys,
           onSelectionChange: setSelectedKeys,
         }}
-        rowConfig={{
-          onClick: item => onSelectConversation(item.id),
-        }}
-        emptyState={
-          <div className={classes.emptyState}>
-            <Text variant="body-medium">No conversations found</Text>
-          </div>
-        }
-      />
-      <DeleteConversationDialog
-        conversation={pendingDelete}
-        onConfirm={() => {
-          if (pendingDelete) deleteConversation(pendingDelete.id);
-          setPendingDelete(null);
-        }}
-        onCancel={() => setPendingDelete(null)}
-      />
-      <RenameConversationDialog
-        conversation={pendingRename}
-        onConfirm={title => {
-          if (pendingRename) renameConversation(pendingRename.id, title);
-          setPendingRename(null);
-        }}
-        onCancel={() => setPendingRename(null)}
       />
       <DeleteConversationsDialog
         count={pendingBulkDelete?.length ?? null}
