+++ plugins/ai-chat/src/components/ConversationHistoryTable/index.ts
@@ -1,0 +1,1 @@
+export { ConversationHistoryTable } from './ConversationHistoryTable';
